gbdelete will no longer work for 'master' or 'develop' branch.

+++ git-aliases.ps1
@@ -34,7 +34,7 @@ function git-branchtrack {
 	}
 	git fetch origin
 	git branch --track $branch_to_track origin/$branch_to_track
-	
+	git checkout $branch_to_track
 }
 Set-Alias -Name gbtrack -Value git-branchtrack
 
@@ -43,9 +43,13 @@ function git-branchdelete {
 	if($branch_to_delete -eq "") {
 		return
 	}
+	if($branch_to_delete -eq "develop" -or $branch_to_delete -eq "master") {
+		write "Cannot delete branch '$branch_to_delete'!"
+		write "If you really want to do this, run the git commands yourself."
+		return
+	}
 	git push origin :refs/heads/$branch_to_delete
 	git branch -d $branch_to_delete
-	
 }
 Set-Alias -Name gbdelete -Value git-branchdelete
